add error handling and ability to change shadow strength and sun angle in renderDEM() as well as make the script re-runnable if opened in GUI. Also fix bugs related to the error handling in fetchDEM() and simplifyDEM()

+++ BlenderMapDEM.py
@@ -4,7 +4,7 @@ import requests
 import os
 import re
 
-# Fetch DEM from OpenTopography of user specified extent
+# Fetch DEM .GeoTIFF image of user specified extent
 def fetchDEM(north_bound: float, south_bound: float, east_bound: float, west_bound: float, API_Key: str, output_dir: str, dataset: str = 'SRTMGL1'):
     # Declare possible DEM datasets
     possible_datasets = ['SRTMGL3',
@@ -22,16 +22,17 @@ def fetchDEM(north_bound: float, south_bound: float, east_bound: float, west_bou
     ### --- Catch a variety of user-input errors --- ###
     
     # Check for invalid characters in output directory
-    pattern = re.compile(r'[^a-zA-Z0-9_\-\\/.\s]')
+    pattern = re.compile(r'[^a-zA-Z0-9_\-\\/.\s:]')
     if pattern.search(output_dir):
         raise ValueError('Input or output directory contains invalid characters.')
     
     # Check for invalid output directory or filetype errors
+    output_dir_path = os.path.dirname(output_dir)
+    if not os.path.exists(output_dir_path):
+        raise ValueError(f'Output file path "{output_dir}" does not exist, please create it.')
     if not output_dir.endswith(('.tif','.tiff')):
         raise ValueError(f'Invalid output filetype "{output_dir}", make sure output_dir argument ends with ".tif"')
-    if not os.path.exists(output_dir):
-        raise ValueError(f'Output file path "{output_dir}" does not exist, please create it.')
-        
+    
     # Raise error if dataset specified by user is not one of the available DEM datasets offered by OpenTopography
     if dataset not in possible_datasets:
         raise ValueError(f'Invalid dataset: "{dataset}" not present in available datasets offered by OpenTopography, see documentation for a list of available datasets')
@@ -77,7 +78,28 @@ def fetchDEM(north_bound: float, south_bound: float, east_bound: float, west_bou
             raise error
 
 # Generate 3D elevation map using Blender
-def renderDEM(dem_dir: str, output_dir: str, exaggeration: float = 0.5, resolution_scale: int = 50, samples: int = 5):     
+def renderDEM(dem_dir: str, output_dir: str, exaggeration: float = 0.5, shadow_softness: float = 90, sun_angle: float = 45, resolution_scale: int = 50, samples: int = 5):
+    
+        ### --- Catch a variety of user-input errors --- ###
+    
+    # Check for invalid characters in input and output directories
+    pattern = re.compile(r'[^a-zA-Z0-9_\-\\/.\s:]')
+    if pattern.search(dem_dir) or pattern.search(output_dir):
+        raise ValueError('Input or output directory contains invalid characters.')
+    
+    # Check for invalid input directory or filetype errors
+    if not os.path.exists(dem_dir):
+        raise ValueError(f'Input file path "{dem_dir}" does not exist.')
+    if not dem_dir.endswith(('.png', '.jpg', '.jpeg', '.bmp','.tif','.tiff')):
+        raise ValueError(f'Input file "{dem_dir}"" is not a valid image file.')
+    
+    # Check for invalid output directory or filetype errors
+    output_dir_path = os.path.dirname(output_dir)
+    if not os.path.exists(output_dir_path):
+        raise ValueError(f'Output file path "{output_dir}" does not exist, please create it.')
+    if not output_dir.endswith(('.png', '.jpg', '.jpeg', '.bmp','.tif','.tiff')):
+        raise ValueError(f'Output file "{output_dir}" is not a valid image file.')  
+    
     # Import DEM image
     DEM = bpy.data.images.load(dem_dir)
     
@@ -99,25 +121,37 @@ def renderDEM(dem_dir: str, output_dir: str, exaggeration: float = 0.5, resoluti
     bpy.data.scenes['Scene'].cycles.samples = samples
     
         ### --- Plane Settings --- ###
+
+    # Get the cube and plane objects if they exists
+    cube = bpy.data.objects.get('Cube')
+    plane = bpy.data.objects.get('Plane')
     
-    # Select and delete the default cube
-    bpy.data.objects['Cube'].select_set(True)
-    bpy.ops.object.delete() 
+    # If the cube exists, delete it
+    if cube != None:
+        bpy.data.objects['Cube'].select_set(True)
+        bpy.ops.object.delete() 
+    
+    # If the cube exists, delete it
+    if plane != None:
+        bpy.data.objects['Plane'].select_set(True)
+        bpy.ops.object.delete() 
 
     # Add plane to scene
     bpy.ops.mesh.primitive_plane_add()
-    
     plane = bpy.data.objects['Plane']
 
     # Scale plane to aspect ratio of image
     plane.scale[0] = width/1000
     plane.scale[1] = height/1000
 
-    # Add subdivision surface modifier to plane
-    bpy.ops.object.modifier_add(type='SUBSURF')
-    bpy.context.object.modifiers["Subdivision"].subdivision_type = 'SIMPLE'
-    bpy.context.object.cycles.use_adaptive_subdivision = True
-    
+    # Check if the subsurf modifier is already applied to the plane
+    subsurf_modifier = plane.modifiers.get('Subdivision')
+    if subsurf_modifier == None:
+        # Add subdivision surface modifier to plane if it doesn't exist
+        subsurf_modifier = plane.modifiers.new(name='Subdivision', type='SUBSURF')
+        subsurf_modifier.subdivision_type = 'SIMPLE'
+        bpy.context.object.cycles.use_adaptive_subdivision = True
+
         ### --- Camera Settings --- ###
     
     camera = bpy.data.objects["Camera"]
@@ -142,23 +176,29 @@ def renderDEM(dem_dir: str, output_dir: str, exaggeration: float = 0.5, resoluti
         orthographic_scale = 2*(height/1000)
     elif height/1000 == width/1000:
         orthographic_scale = 2*(height/1000)
-
+    
     bpy.data.cameras['Camera'].ortho_scale = orthographic_scale
     
         ### --- Light Settings --- ###
     
     light = bpy.data.objects['Light']
     
+    # Make sure shadow_strength falls in acceptable range
+    if shadow_softness > 180:
+        shadow_softness = 180
+    elif shadow_softness < 0:
+        shadow_softness = 0
+    
     # Select light and change its type to "Sun"
     bpy.data.lights['Light'].type = 'SUN'
     
     # Change strength and hardness of light
     bpy.data.lights['Light'].energy = 5
-    bpy.data.lights['Light'].angle = 1.5708
+    bpy.data.lights['Light'].angle = shadow_softness/57.295
     
     # Change direction of the light
     light.rotation_euler[0] = 0
-    light.rotation_euler[1] = 0.785398
+    light.rotation_euler[1] = sun_angle/57.295
     light.rotation_euler[2] = 2.35619
     
         ### --- Shader Settings --- ###
@@ -174,19 +214,30 @@ def renderDEM(dem_dir: str, output_dir: str, exaggeration: float = 0.5, resoluti
     mat.node_tree.nodes['Principled BSDF'].inputs['Roughness'].default_value = 1
     mat.node_tree.nodes['Principled BSDF'].inputs['Specular'].default_value = 0
     
-    # Add image texture node and specify its extension, interpolation, and colorspace modes
-    imageTexture = mat.node_tree.nodes.new('ShaderNodeTexImage')
-    imageTexture.image = DEM
-    mat.node_tree.nodes["Image Texture"].extension = 'EXTEND'
-    mat.node_tree.nodes["Image Texture"].interpolation = 'Smart'
-    mat.node_tree.nodes['Image Texture'].image.colorspace_settings.name='Linear'
+    # Add image texture node if it doesnt exist and specify its extension, interpolation, and colorspace modes
+    if 'Image Texture' not in mat.node_tree.nodes:
+        imageTexture = mat.node_tree.nodes.new('ShaderNodeTexImage')
+    else:
+        imageTexture = mat.node_tree.nodes['Image Texture']
 
-    # Add displacement node and specify the elevation exaggeration of heightmap
-    displacement = mat.node_tree.nodes.new('ShaderNodeDisplacement')
-    mat.node_tree.nodes['Displacement'].inputs['Scale'].default_value = exaggeration
+    imageTexture.image = DEM
+    imageTexture.extension = 'EXTEND'
+    imageTexture.interpolation = 'Smart'
+    imageTexture.image.colorspace_settings.name = 'Linear'
     
-    # Add color ramp node
-    colorRamp = mat.node_tree.nodes.new('ShaderNodeValToRGB')
+    # Add displacement node if it doesnt exist and specify the elevation exaggeration of heightmap
+    if 'Displacement' not in mat.node_tree.nodes:
+        displacement = mat.node_tree.nodes.new('ShaderNodeDisplacement')
+    else:
+        displacement = mat.node_tree.nodes['Displacement']
+        
+    displacement.inputs['Scale'].default_value = exaggeration
+    
+    # Add color ramp node if it doesnt exist
+    if 'ColorRamp' not in mat.node_tree.nodes:
+        colorRamp = mat.node_tree.nodes.new('ShaderNodeValToRGB')
+    else:
+        colorRamp = mat.node_tree.nodes['ColorRamp']
     
     # Link everything together
     mat.node_tree.links.new(imageTexture.outputs['Color'], displacement.inputs['Height'])
@@ -194,9 +245,19 @@ def renderDEM(dem_dir: str, output_dir: str, exaggeration: float = 0.5, resoluti
        
         ### --- Render Image --- ###
     
+    # Check output file type and set accordingly
+    if output_dir.endswith(('.tif','.tiff')):
+        file_type = 'TIFF'
+    elif output_dir.endswith(('.jpg', '.jpeg')):
+        file_type = 'JPEG'
+    elif output_dir.endswith(('.bmp')):
+        file_type = 'BMP'
+    else:
+        file_type = 'PNG'
+    
     # Set the output file path and format
     bpy.context.scene.render.filepath = output_dir
-    bpy.context.scene.render.image_settings.file_format = 'PNG'
+    bpy.context.scene.render.image_settings.file_format = file_type
     
     # Render the image
     bpy.ops.render.render(write_still=True)
@@ -207,21 +268,22 @@ def simplifyDEM(dem_dir: str, output_dir: str, reduction_factor: int = 2):
     ### --- Catch a variety of user-input errors --- ###
     
     # Check for invalid characters in input and output directories
-    pattern = re.compile(r'[^a-zA-Z0-9_\-\\/.\s]')
+    pattern = re.compile(r'[^a-zA-Z0-9_\-\\/.\s:]')
     if pattern.search(dem_dir) or pattern.search(output_dir):
         raise ValueError('Input or output directory contains invalid characters.')
     
     # Check for invalid input directory or filetype errors
     if not os.path.exists(dem_dir):
         raise ValueError(f'Input file path "{dem_dir}" does not exist.')
-    if not dem_dir.endswith(('.png', '.jpg', '.jpeg', '.bmp','.tif','.tiff')):
+    if not dem_dir.endswith(('.png','.bmp','.tif','.tiff')):
         raise ValueError(f'Input file "{dem_dir}"" is not a valid image file.')
     
     # Check for invalid output directory or filetype errors
-    if not output_dir.endswith(('.png', '.jpg', '.jpeg', '.bmp','.tif','.tiff')):
-        raise ValueError(f'Output file "{dem_dir}" is not a valid image file.')  
-    if not os.path.exists(output_dir):
+    output_dir_path = os.path.dirname(output_dir)
+    if not os.path.exists(output_dir_path):
         raise ValueError(f'Output file path "{output_dir}" does not exist, please create it.')
+    if not output_dir.endswith(('.png','.bmp','.tif','.tiff')):
+        raise ValueError(f'Output file "{output_dir}" is not a valid image file.')  
     
     # Check for invalid reduction_factor that would result in the same or larger image
     if reduction_factor < 2:
@@ -242,9 +304,3 @@ def simplifyDEM(dem_dir: str, output_dir: str, reduction_factor: int = 2):
 
     # Save the downscaled image to a new file
     simplified_img.save(output_dir)
-
-# Fetch satellite imagery of user specified extent
-#def fetchImagery(upper_lat,lower_lat,left_lon,right_lon, output_dir, filename = 'imagery.tif'):
-
-# Plots 2D DEM visualization 
-#def plotDEM(dem)
